requirements.txt
requests
opencv-python==3.4.0.12
paho-mqtt